use more strict casts

+++ src/project.cpp
@@ -45,7 +45,7 @@ void inputThread() {
       int mouseX = GetMouseX();
       int mouseY = GetMouseY();
 
-      if (CheckCollisionPointRec((Vector2){mouseX, mouseY}, buttonRec)) {
+      if (CheckCollisionPointRec((Vector2){static_cast<float>(mouseX), static_cast<float>(mouseY)}, buttonRec)) {
         continue;
       }
 
@@ -64,7 +64,7 @@ void inputThread() {
       int mouseX = GetMouseX();
       int mouseY = GetMouseY();
 
-      if (CheckCollisionPointRec((Vector2){mouseX, mouseY}, buttonRec)) {
+      if (CheckCollisionPointRec((Vector2){static_cast<float>(mouseX), static_cast<float>(mouseY)}, buttonRec)) {
         continue;
       }
